requirements_docker.txt
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
fastapi==0.115.4
pinecone==5.3.1
pydantic-settings==2.6.1
spacy==3.8.2
transformers==4.46.2
uvicorn==0.32.0
sentence-transformers
